requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
cmake==3.30.0
contourpy==1.1.1
cycler==0.12.1
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
idna==3.7
importlib_resources==6.4.0
imutils==0.5.4
Jinja2==3.1.4
kiwisolver==1.4.5
lightning==2.3.3
lightning-utilities==0.11.4
lit==18.1.8
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
packaging==24.1
pandas==2.0.3
pillow==10.4.0
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.1.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scipy==1.10.1
six==1.16.0
sympy==1.13.0
torch==2.0.1
torchmetrics==1.4.0.post0
torchvision==0.15.2
tqdm==4.65.2
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
zipp==3.19.2
